src/requirements.txt
requests
mysql-connector-python
dotenv
openai